Moved router config to a separate file
