+++ src/main.js
@@ -1,47 +1,9 @@
 import Vue from 'vue'
 import App from './App.vue'
 
-import VueRouter from 'vue-router';
-
-import Home from './pages/home'
-import Balance from './pages/balance'
-import Cards from './pages/cards'
-import Account from './pages/account'
-import Settings from './pages/settings'
+import router from './router'
 
 import '../node_modules/normalize.css/normalize.css';
-
-// Router configuration
-Vue.use(VueRouter)
-
-const routes = [
-  {
-    path: '/',
-    component: Home
-  },
-  {
-    path: '/balance',
-    component: Balance
-  },
-  {
-    path: '/cards',
-    component: Cards
-  },
-  {
-    path: '/account',
-    component: Account
-  },
-  {
-    path: '/settings',
-    component: Settings
-  }
-]
-
-const router = new VueRouter({
-  mode: 'history',
-  linkActiveClass: '-active',
-  routes
-})
 
 // Vue configuration
 new Vue({
